extract UserRoles concern from User model
